RLP-865: fragment matrix/transport.py (#152)
fragment matrix/transport.py into node.py and utils.py
