+++ raiden/tests/integration/fixtures/transport.py
@@ -3,7 +3,7 @@ import pytest
 from raiden.constants import DISCOVERY_DEFAULT_ROOM, PATH_FINDING_BROADCASTING_ROOM
 from raiden.tests.fixtures.variables import TransportProtocol
 from raiden.tests.utils.transport import generate_synapse_config, matrix_server_starter
-from transport.matrix.transport import MatrixNode as MatrixTransportNode
+from transport.matrix.node import MatrixNode as MatrixTransportNode
 
 
 @pytest.fixture
