+++ raiden/tests/integration/network/transport/test_matrix_transport.py
@@ -22,8 +22,8 @@ from raiden.transfer.state_change import ActionChannelClose, ActionUpdateTranspo
 from raiden.utils.signer import LocalSigner
 from raiden.utils.typing import Address, List, Optional, Union
 from transport.matrix.client import Room
-from transport.matrix.transport import MatrixNode as MatrixTransportNode, _RetryQueue
-from transport.matrix.utils import AddressReachability, make_room_alias
+from transport.matrix.node import MatrixNode as MatrixTransportNode
+from transport.matrix.utils import AddressReachability, make_room_alias, _RetryQueue
 from transport.message import Message as TransportMessage
 
 USERID0 = "@Arthur:RestaurantAtTheEndOfTheUniverse"
@@ -169,12 +169,12 @@ def skip_userid_validation(monkeypatch):
         return factories.HOP1
 
     monkeypatch.setattr(
-        transport.matrix.transport,
+        transport.matrix.node,
         "validate_userid_signature",
         mock_validate_userid_signature,
     )
     monkeypatch.setattr(
-        transport.matrix.utils,
+        transport.matrix.node,
         "validate_userid_signature",
         mock_validate_userid_signature,
     )
