+++ raiden/tests/integration/test_send_queued_messages.py
@@ -18,7 +18,7 @@ from raiden.transfer import views
 from raiden.transfer.events import EventPaymentSentSuccess
 from raiden.transfer.mediated_transfer.events import SendSecretReveal
 from raiden.utils import BlockNumber
-from transport.matrix.transport import MatrixNode as MatrixTransportNode
+from transport.matrix.node import MatrixNode as MatrixTransportNode
 
 
 @pytest.mark.parametrize("deposit", [10])
